add trainer base class
