+++ TODO.md
@@ -1,4 +1,5 @@
 ## 2022-11-21
-- [ ] train_dataset, test_dataset, behavior when either is none?
-- [ ] deprecate build_model function 
-- [ ] 
+- [X] train_dataset, test_dataset, behavior when either is none?
+- [X] deprecate build_model function. Models each take in a separate config. Still used!
+- [X] Base trainer.py.
+- [ ] torchinfo dependency
